+++ service/src/main/java/com/cisco/prepos/services/discount/DiscountProvider.java
@@ -15,6 +15,6 @@ public interface DiscountProvider {
 
     double getDiscount(Dart selectedDart, Promo promo, Pricelist pricelist, long shippedDateInMillis);
 
-    int getGpl(String partNumber, Map<String, Pricelist> pricelistsMap);
+    double getGpl(String partNumber, Map<String, Pricelist> pricelistsMap);
 
 }
