+++ service/src/test/java/com/cisco/prepos/services/recount/DefaultPreposRecounterTest.java
@@ -43,7 +43,7 @@ public class DefaultPreposRecounterTest {
 
     private static final String SELECTED_DART_AUTHORIZATION_NUMBER = "other authorization number";
     private static final String SELECTED_DART_END_USER_NAME = "other end user name";
-    private static final int OTHER_GPL = 270;
+    private static final Double OTHER_GPL = 270d;
 
     @InjectMocks
     private DefaultPreposRecounter defaultPreposRecounter = new DefaultPreposRecounter();
