+++ service/src/test/resources/sql/hsqldb/005_pricelist.sql
@@ -2,7 +2,7 @@ CREATE TABLE pricelist (
   id        BIGINT IDENTITY NOT NULL PRIMARY KEY,
   part_number VARCHAR(1000)      NOT NULL,
   description      VARCHAR(1000)      NOT NULL,
-  gpl       INT              NOT NULL,
+  gpl       DOUBLE              NOT NULL,
   wpl       DOUBLE,
   discount    DOUBLE
 );
